chore: delete redudant code
